Debugging. Style fixes.
- Fix pyflakes errors
- Debugging cause of Sundials not using preconditioner
- Update README.rst

+++ src/chemreac.hpp
@@ -65,11 +65,11 @@ private:
 
 public:
     double * xc; // bin centers (length = N+nstencil-1), first bin center: xc[(nstencil-1)/2]
-    uint neval_f {0};
-    uint neval_j {0};
-    uint nprec_setup {0};
-    uint nprec_solve {0};
-    uint njacvec_dot {0};
+    long neval_f {0};
+    long neval_j {0};
+    long nprec_setup {0};
+    long nprec_solve {0};
+    long njacvec_dot {0};
 
     ReactionDiffusion(uint, 
 		      const vector<vector<uint> >,
